# docs/source/reference/api-restapi/openapi.yml
openapi: 3.0.1
info:
  title: Radar REST API
  version: 0.0.0
servers:
- url: /
...
paths:
...
  /api/v1/entrypoints/{id}/tags:
    get:
      tags:
      - Entrypoints
      summary: Gets the list of Tags for the resource
      operationId: get_resources_tags_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the entry_point.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagRef-dump'
    put:
      tags:
      - Entrypoints
      summary: Replaces one or more Tags to the resource
      operationId: put_resources_tags_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the entry_point.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdList'
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagRef-dump'
      x-codegen-request-body-name: payload
    post:
      tags:
      - Entrypoints
      summary: Appends one or more Tags to the resource
      operationId: post_resources_tags_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the entry_point.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdList'
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagRef-dump'
      x-codegen-request-body-name: payload
    delete:
      tags:
      - Entrypoints
      summary: Removes all Tags from the resource
      operationId: delete_resources_tags_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the entry_point.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdStatusResponse'
  /api/v1/entrypoints/{id}/tags/{tagId}:
    delete:
      tags:
      - Entrypoints
      summary: Removes a Tag from the resource
      operationId: delete_resources_tags_id_endpoint
      parameters:
      - name: tagId
        in: path
        description: ID for the Tag.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the entry_point.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdStatusResponse'
...
  /api/v1/experiments/{id}/tags:
    get:
      tags:
      - Experiments
      summary: Gets the list of Tags for the resource
      operationId: get_resources_tags_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the experiment.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagRef-dump'
    put:
      tags:
      - Experiments
      summary: Replaces one or more Tags to the resource
      operationId: put_resources_tags_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the experiment.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdList'
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagRef-dump'
      x-codegen-request-body-name: payload
    post:
      tags:
      - Experiments
      summary: Appends one or more Tags to the resource
      operationId: post_resources_tags_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the experiment.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdList'
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagRef-dump'
      x-codegen-request-body-name: payload
    delete:
      tags:
      - Experiments
      summary: Removes all Tags from the resource
      operationId: delete_resources_tags_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the experiment.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdStatusResponse'
  /api/v1/experiments/{id}/tags/{tagId}:
    delete:
      tags:
      - Experiments
      summary: Removes a Tag from the resource
      operationId: delete_resources_tags_id_endpoint
      parameters:
      - name: tagId
        in: path
        description: ID for the Tag.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the experiment.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdStatusResponse'
...
  /api/v1/jobs/{id}/tags:
    get:
      tags:
      - Job
      summary: Gets the list of Tags for the resource
      operationId: get_resources_tags_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the job.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagRef-dump'
    put:
      tags:
      - Job
      summary: Replaces one or more Tags to the resource
      operationId: put_resources_tags_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the job.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdList'
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagRef-dump'
      x-codegen-request-body-name: payload
    post:
      tags:
      - Job
      summary: Appends one or more Tags to the resource
      operationId: post_resources_tags_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the job.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdList'
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagRef-dump'
      x-codegen-request-body-name: payload
    delete:
      tags:
      - Job
      summary: Removes all Tags from the resource
      operationId: delete_resources_tags_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the job.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdStatusResponse'
  /api/v1/jobs/{id}/tags/{tagId}:
    delete:
      tags:
      - Job
      summary: Removes a Tag from the resource
      operationId: delete_resources_tags_id_endpoint
      parameters:
      - name: tagId
        in: path
        description: ID for the Tag.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the job.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdStatusResponse'
...
  /api/v1/models/{id}/tags:
    get:
      tags:
      - Models
      summary: Gets the list of Tags for the resource
      operationId: get_resources_tags_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the ml_model.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagRef-dump'
    put:
      tags:
      - Models
      summary: Replaces one or more Tags to the resource
      operationId: put_resources_tags_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the ml_model.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdList'
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagRef-dump'
      x-codegen-request-body-name: payload
    post:
      tags:
      - Models
      summary: Appends one or more Tags to the resource
      operationId: post_resources_tags_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the ml_model.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdList'
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagRef-dump'
      x-codegen-request-body-name: payload
    delete:
      tags:
      - Models
      summary: Removes all Tags from the resource
      operationId: delete_resources_tags_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the ml_model.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdStatusResponse'
  /api/v1/models/{id}/tags/{tagId}:
    delete:
      tags:
      - Models
      summary: Removes a Tag from the resource
      operationId: delete_resources_tags_id_endpoint
      parameters:
      - name: tagId
        in: path
        description: ID for the Tag.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the ml_model.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdStatusResponse'
...
  /api/v1/pluginParameterTypes/{id}/tags:
    get:
      tags:
      - PluginParameterTypes
      summary: Gets the list of Tags for the resource
      operationId: get_resources_tags_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the plugin_task_parameter_type.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagRef-dump'
    put:
      tags:
      - PluginParameterTypes
      summary: Replaces one or more Tags to the resource
      operationId: put_resources_tags_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the plugin_task_parameter_type.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdList'
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagRef-dump'
      x-codegen-request-body-name: payload
    post:
      tags:
      - PluginParameterTypes
      summary: Appends one or more Tags to the resource
      operationId: post_resources_tags_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the plugin_task_parameter_type.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdList'
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagRef-dump'
      x-codegen-request-body-name: payload
    delete:
      tags:
      - PluginParameterTypes
      summary: Removes all Tags from the resource
      operationId: delete_resources_tags_endpoint
      parameters:
      - name: id
        in: path
        description: ID for the plugin_task_parameter_type.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdStatusResponse'
  /api/v1/pluginParameterTypes/{id}/tags/{tagId}:
    delete:
      tags:
      - PluginParameterTypes
      summary: Removes a Tag from the resource
      operationId: delete_resources_tags_id_endpoint
      parameters:
      - name: tagId
        in: path
        description: ID for the Tag.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the plugin_task_parameter_type.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdStatusResponse'
...
  /api/v1/plugins/{id}/files/{fileId}/tags/:
    get:
      tags:
      - Plugins
      summary: Gets the list of Tags for the resource
      operationId: get_resources_tags_endpoint
      parameters:
      - name: fileId
        in: path
        description: ID for the plugin_file.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the plugin.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagRef-dump'
    put:
      tags:
      - Plugins
      summary: Replaces one or more Tags to the resource
      operationId: put_resources_tags_endpoint
      parameters:
      - name: fileId
        in: path
        description: ID for the plugin_file.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the plugin.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdList'
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagRef-dump'
      x-codegen-request-body-name: payload
    post:
      tags:
      - Plugins
      summary: Appends one or more Tags to the resource
      operationId: post_resources_tags_endpoint
      parameters:
      - name: fileId
        in: path
        description: ID for the plugin_file.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the plugin.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdList'
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagRef-dump'
      x-codegen-request-body-name: payload
    delete:
      tags:
      - Plugins
      summary: Removes all Tags from the resource
      operationId: delete_resources_tags_endpoint
      parameters:
      - name: fileId
        in: path
        description: ID for the plugin_file.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the plugin.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdStatusResponse'
  /api/v1/plugins/{id}/files/{fileId}/tags/{tagId}:
    delete:
      tags:
      - Plugins
      summary: Removes a Tag from the resource
      operationId: delete_resources_tags_id_endpoint
      parameters:
      - name: tagId
        in: path
        description: ID for the Tag.
        required: true
        schema:
          type: integer
      - name: fileId
        in: path
        description: ID for the plugin_file.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the plugin.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdStatusResponse'
...
  /api/v1/plugins/{id}/tags/{tagId}:
    delete:
      tags:
      - Plugins
      summary: Removes a Tag from the resource
      operationId: delete_resources_tags_id_endpoint
      parameters:
      - name: tagId
        in: path
        description: ID for the Tag.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the plugin.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdStatusResponse'
...
  /api/v1/queues/{id}/tags/{tagId}:
    delete:
      tags:
      - Queues
      summary: Removes a Tag from the resource
      operationId: delete_resources_tags_id_endpoint
      parameters:
      - name: tagId
        in: path
        description: ID for the Tag.
        required: true
        schema:
          type: integer
      - name: id
        in: path
        description: ID for the queue.
        required: true
        schema:
          type: integer
      - name: X-Fields
        in: header
        description: An optional fields mask
        schema:
          type: string
          format: mask
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdStatusResponse'
...
components:
  schemas:
...
    GroupRef:
      type: object
      properties:
        id:
          type: integer
          description: ID for the Group resource.
        name:
          type: string
          description: Name of the Group resource.
        url:
          type: string
          description: URL for accessing the full Group resource.
...
    TagRef-dump:
      type: object
      properties:
        id:
          type: integer
          description: ID for the Tag.
        name:
          type: string
          description: Name of the Tag.
        group:
          type: object
          description: Owner of the Tag.
          allOf:
          - $ref: '#/components/schemas/GroupRef'
        url:
          type: string
          description: URL for accessing the full Tag.
...
    IdStatusResponse:
      type: object
      properties:
        id:
          type: array
          description: A list of integers identifying the affected resource(s).
          items:
            type: integer
        status:
          type: string
          description: The status of the request.
...
    IdList:
      type: object
      properties:
        ids:
          type: array
          description: List of identifiers for one or more objects.
          items:
            type: integer
...